Dirty workaround for numeric IDs as relationship targets
Without this, autocompletion in relationship field causes an error, as
it is trying to feed a string in GraphQL `{id: {equals: ...}}`.

This is *not* the right fix, as there could be numeric fields that are
not autoincrement, but at the time of this writing, the necessary
information about the field type does not seem to be available to this
function.

See https://github.com/keystonejs/keystone/issues/8784